Rename Cyclic Permutation to Cyclic Permutation Matrix
