+++ toqito/matrices/__init__.py
@@ -10,4 +10,4 @@ from toqito.matrices.gen_pauli import gen_pauli
 from toqito.matrices.hadamard import hadamard
 from toqito.matrices.pauli import pauli
 from toqito.matrices.standard_basis import standard_basis
-from toqito.matrices.cyclic_permutation import cyclic_permutation_matrix
+from toqito.matrices.cyclic_permutation_matrix import cyclic_permutation_matrix
